Refactoring common macros
This patch refactors assertion-related and ASIO concurrent macros.
